renamed and refactred and packages and classes for better structure and navigation
