+++ src/main/java/vaultiq/session/cache/service/internal/BlocklistSessionCacheService.java
@@ -1,4 +1,4 @@
-package vaultiq.session.cache.service;
+package vaultiq.session.cache.service.internal;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,12 +7,12 @@ import org.springframework.cache.Cache;
 import org.springframework.stereotype.Service;
 import vaultiq.session.cache.model.ModelType;
 import vaultiq.session.cache.model.SessionIds;
-import vaultiq.session.cache.utility.VaultiqCacheContext;
-import vaultiq.session.core.VaultiqSessionContext;
+import vaultiq.session.cache.util.VaultiqCacheContext;
+import vaultiq.session.core.util.VaultiqSessionContext;
 
 import java.util.Optional;
 
-import static vaultiq.session.cache.utility.CacheKeyResolver.keyForBlacklist;
+import static vaultiq.session.cache.util.CacheKeyResolver.keyForBlacklist;
 
 @Service
 @ConditionalOnBean(VaultiqCacheContext.class)
