+++ src/main/java/vaultiq/session/cache/util/SessionIdRequestMapper.java
@@ -1,11 +1,11 @@
-package vaultiq.session.utility;
+package vaultiq.session.cache.util;
 
 import jakarta.servlet.http.HttpServletRequest;
 
-public class VaultiqRequestContext {
+public class SessionIdRequestMapper {
     private static final String VAULTIQ_SID_KEY = "vaultiq.sid";
 
-    public static void setSessionId(HttpServletRequest request, String sessionId) {
+    public static void mapSessionId(HttpServletRequest request, String sessionId) {
         request.setAttribute(VAULTIQ_SID_KEY, sessionId);
     }
