+++ src/main/java/vaultiq/session/cache/util/VaultiqCacheContext.java
@@ -1,4 +1,4 @@
-package vaultiq.session.cache.utility;
+package vaultiq.session.cache.util;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
