+++ src/main/java/vaultiq/session/config/VaultiqBeanCondition.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Condition;
 import org.springframework.context.annotation.ConditionContext;
 import org.springframework.core.type.AnnotatedTypeMetadata;
 import vaultiq.session.cache.model.ModelType;
-import vaultiq.session.core.VaultiqSessionContext;
+import vaultiq.session.core.util.VaultiqSessionContext;
 
 import java.util.Map;
 import java.util.Objects;
