+++ src/main/java/vaultiq/session/core/util/VaultiqSessionContext.java
@@ -1,10 +1,9 @@
-package vaultiq.session.core;
+package vaultiq.session.core.util;
 
 import org.springframework.stereotype.Component;
 import vaultiq.session.cache.model.ModelType;
 import vaultiq.session.config.VaultiqSessionProperties;
 import vaultiq.session.core.model.VaultiqModelConfig;
-import vaultiq.session.core.util.VaultiqModelConfigEnhancer;
 
 import java.util.Map;
